requirements.txt
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
# dvc==3.48.4
# dvc-gdrive==3.0.1
PyYAML==6.0.1
pandas==2.2.1
matplotlib==3.8.3
pillow==10.2.0
scikit-learn==1.3.2
transformers==4.38.2
opencv-python==4.9.0.80
tyro==0.8.3
albumentations
lightning
pytorch_optimizer
pytorch_lightning